abc: Enable usage of rmp library for custom implementations.
